+++ changelog.md
@@ -29,3 +29,4 @@
 - [unitmodules.cpp/hpp] DELETED: 因为没找到合适的算法，所以暂时取消删除多余模块的功能.
 - [baseclass.hpp] DELETED: `Set_Enable`.
 - CHANGED: 所有应该使用uint的int类型改为uint。
+- [simulator.cpp/hpp] CHANGED: `_outref`改为数组.